# backend-go/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateOfficeRequest:
    properties:
      address:
        type: string
      is_active:
        type: boolean
      office_name:
        type: string
      office_type:
        type: string
    required:
    - address
    - office_name
    - office_type
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
    type: object
...
  dto.UpdateOfficeRequest:
    properties:
      address:
        type: string
      is_active:
        type: boolean
      office_name:
        type: string
      office_type:
        type: string
    type: object
...
  entity.Office:
    properties:
      address:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      office_name:
        type: string
      office_type:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@ev-warranty.com
    name: API Support
  description: API for EV Warranty Management System
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EV Warranty API
  version: "1.0"
paths:
...
  /offices:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all offices
      produces:
      - application/json
      responses:
        "200":
          description: Offices retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Office'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Get all offices
      tags:
      - offices
    post:
      consumes:
      - application/json
      description: Create a new office (Admin only)
      parameters:
      - description: Office creation data
        in: body
        name: createOfficeRequest
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOfficeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Office created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.Office'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new office
      tags:
      - offices
  /offices/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an office by ID (Admin only)
      parameters:
      - description: Office ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Office deleted successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Office not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete an office
      tags:
      - offices
    get:
      consumes:
      - application/json
      description: Retrieve a specific office by its ID
      parameters:
      - description: Office ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Office retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.Office'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Office not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Get office by ID
      tags:
      - offices
    put:
      consumes:
      - application/json
      description: Update an existing office by ID (Admin only)
      parameters:
      - description: Office ID
        in: path
        name: id
        required: true
        type: string
      - description: Office update data
        in: body
        name: updateOfficeRequest
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOfficeRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Office updated successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Office not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Update an office
      tags:
      - offices
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"